+++ R/dim.R
@@ -1,6 +1,6 @@
 #' Vector dimensions
 #'
-#' * `vec_dims()` gives the dimensionality (i.e. number of dimensions)
+#' * `vec_n_dim()` gives the dimensionality (i.e. number of dimensions)
 #' * `vec_dim()` returns the size of each dimension
 #'
 #' Unlike base R, we treat vectors with `NULL` dimensions as 1d. This
@@ -8,7 +8,7 @@
 #' equivalent, `vec_dim()` returns `length()`, not `NULL`, when `x` is 1d.
 #'
 #' @param x A vector
-#' @name dims
+#' @name dim
 #' @examples
 #' # Compared to base R
 #' x <- 1:5
@@ -17,18 +17,19 @@
 NULL
 
 #' @export
-#' @rdname dims
+#' @rdname dim
 vec_dim <- function(x) {
-  dim(x) %||% length(x)
+  .Call(vctrs_dim, x)
 }
 
 #' @export
-#' @rdname dims
-vec_dims <- function(x) {
-  d <- dim(x)
-  if (is.null(d)) {
-    1L
-  } else {
-    length(d)
-  }
+#' @rdname dim
+vec_n_dim <- function(x) {
+  .Call(vctrs_n_dim, x)
+}
+
+# Unexported counterpart to `vec_dim()` that respects that `dim()` is a generic.
+# Data frames looks like 2d structures with this.
+dim2 <- function(x) {
+  dim(x) %||% length(x)
 }